HtmlBridge Phase 3: de-list AnimationResolver.cs under the 750-line ratchet
The complementary ratchet action to the AnchorFunctions split: shrink an *exempt*
file under the limit and remove its exemption. AnimationResolver.cs (760) had its
cohesive CSS timing-function / easing cluster — ApplyTimingFunction, SolveCubicBezier
with BezierCoord/BezierDerivative, ApplySteps, and the two [GeneratedRegex] steps() /
cubic-bezier() patterns — split into the sibling DomBridge partial
AnimationResolver.Timing.cs (127 lines), dropping AnimationResolver.cs to 644.

Pure partial-class relocation: no signature, accessibility, or logic change, so
behaviour-identical by construction. Its entry is removed from the guard's
OversizedFileExemptions, so the ratchet's stale-exemption check now enforces it stays
under 750. Documented debt list: nine -> eight files.

Verified: Broiler.HtmlBridge.Dom builds clean; the line-limit guard passes (0 offenders
/ 0 stale exemptions); the animation test suite (8) passes.
